requirements.txt
qdrant_client
pydantic
fastapi
uvicorn
python-multipart
tqdm
Pillow
numpy
torch
open_clip_torch
transformers
python-dotenv
